+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from "./lib/translations/translations";
 export * from "./lib/useReactDataTableState/useReactDataTableState";
 export * from "./lib/useReactDataTableState/useReactDataTableStateProps";
+export * from "./lib/useReactDataTableState/usePersistentReactDataTableState";
 export * from "./lib/useReactDataTable/useReactDataTable";
 export * from "./lib/useReactDataTable/useReactDataTableProps";
 export * from "./lib/useReactDataTable/useReactDataTableResult";